[ODS-5916] Integrate latest 5.0 Data Standard Work - Final (#113)
* homograph Extension updated

* Sample Extension MetaEd Artifacts updated

* SampleStudentTranscript Extension MetaEd Artifacts Updated

* samplestudenttransportation Extension - MetaEd Artifacts updated

* Tpdm Extension - MetaEd Artifacts Updated

* [DATASRD-2018] - Rename elements to prevent TPDM related conflicts

* Rename elements to prevent TPDM related conflicts

* [METAED-1481] FK reverse indexes removed for subtables only

+++ Ed-Fi-Extensions/Ed-Fi-Extensions/Extensions/EdFi.Ods.Extensions.SampleStudentTransportation/Versions/1.0.0/Standard/5.0.0/Artifacts/MsSql/Structure/Ods/0020-EXTENSION-SampleStudentTransportation-Tables.sql
@@ -7,7 +7,7 @@
 CREATE TABLE [samplestudenttransportation].[StudentTransportation] (
     [AMBusNumber] [NVARCHAR](6) NOT NULL,
     [PMBusNumber] [NVARCHAR](6) NOT NULL,
-    [SchoolId] [INT] NOT NULL,
+    [SchoolId] [BIGINT] NOT NULL,
     [StudentUSI] [INT] NOT NULL,
     [EstimatedMilesFromSchool] [DECIMAL](5, 2) NOT NULL,
     [Discriminator] [NVARCHAR](128) NULL,
